--- a/Kitodo/src/main/java/org/kitodo/production/helper/cache/RequestScopeCacheHelper.java
+++ b/Kitodo/src/main/java/org/kitodo/production/helper/cache/RequestScopeCacheHelper.java
@@ -43,9 +43,6 @@ public class RequestScopeCacheHelper {
 
     private static final String ATTRIBUTE_NAME = "RequestScopeCacheHelper";
 
-    @SuppressWarnings("unused")
-    private static final Logger logger = LogManager.getLogger(RequestScopeCacheHelper.class);
-
     /**
      * Load or create a simple hash map in the FacesContext.getAttributes() map.
      * @return the cache